Fixing compilations.
