--- core/currency/src/androidMain/kotlin/com/adammcneilly/pocketleague/core/currency/CurrencyFormatter.kt
@@ -1,8 +1,0 @@
-package com.adammcneilly.pocketleague.core.currency
-
-/**
- * Returns the [CurrencyFormatter] instance for this target.
- */
-actual fun currencyFormatter(): CurrencyFormatter {
-    return AndroidCurrencyFormatter()
-}
