--- core/datetime/src/androidMain/kotlin/com/adammcneilly/pocketleague/core/datetime/DateTimeFormatter.kt
@@ -1,10 +1,0 @@
-package com.adammcneilly.pocketleague.core.datetime
-
-/**
- * Creates a [AndroidDateTimeFormatter] to be used on the Android platform.
- *
- * This should be moved into a JVM model another time.
- */
-actual fun dateTimeFormatter(): DateTimeFormatter {
-    return AndroidDateTimeFormatter()
-}
